+++ src/widgets/mod.rs
@@ -1,7 +1,11 @@
+pub mod stdoutput;
+
 use iced::{
     widget::{button, container, row, tooltip},
-    Element,
+    Element, Theme,
 };
+
+use crate::{icons, Message};
 
 pub fn target_card<'a, Message: Clone + 'a>(
     action: Element<'a, Message>,
@@ -31,5 +35,29 @@ pub fn action<'a, Message: Clone + 'a>(
         .into()
     } else {
         action.style(button::secondary).into()
+    }
+}
+
+pub fn target_list<'a>(
+    targets: &'a [String],
+    list: &mut Vec<
+        iced_core::Element<
+            'a,
+            Message,
+            Theme,
+            iced_renderer::fallback::Renderer<iced_wgpu::Renderer, iced_tiny_skia::Renderer>,
+        >,
+    >,
+) {
+    for (id, target) in targets.iter().enumerate() {
+        list.push(target_card(
+            action(
+                icons::start_icon(),
+                target,
+                Some(Message::TaskMake(id, target.clone())),
+            ),
+            target,
+            action(icons::stop_icon(), "stop", Some(Message::TaskStop(id))),
+        ));
     }
 }