+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+[v0.13.0] 2024-12-06
+-------------------
+
+-   refactor StdOutput to widget
+-   targets moved to widget
+-   update README.md
+
 [v0.12.0] 2024-11-27
 -------------------